requirements.txt
amqp==5.1.1
asgiref==3.6.0
autopep8==2.0.1
billiard==3.6.4.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-filter==22.1
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-yasg==1.21.4
flower==1.2.0
humanize==4.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.10.0
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
shortuuid==1.0.11
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
